research(v6): phase 1a fixes — bf16 dtype + China mirrors in bootstrap
Two fixes needed to get Phase 1a to green on real hardware:

1) dtype=bfloat16 (was float16)
   fp16 triggered NaN in Qwen2.5 attention softmax on some prompts when
   attention scores approached fp16's 65504 max. Qwen2.5 is natively
   bfloat16 trained — running fp16 is a post-hoc choice that broke
   numerics on one of four test prompts ('!!!!!!!!!!' garbage output
   with NaN logits). bf16 shares fp32's exponent range so no overflow.
   Trade: reduced mantissa precision, but determinism work doesn't
   require mantissa precision, only reproducibility.

   Knock-on fix: torch.bfloat16 tensors can't .numpy() directly
   (numpy has no native bf16). Added .float() upcast before .numpy()
   in both WorkerSimulator and ReplayEngine. Upcast is lossless
   (bf16 → fp32 just widens the exponent).

2) Bootstrap overrides for China mainland egress
   download.pytorch.org is hard-rate-limited (~14 kB/s) from Aliyun
   mainland. Added PIP_INDEX_URL + TORCH_FIND_LINKS env vars so the
   operator can redirect to a local mirror (e.g. mirrors.aliyun.com).
   TORCH_VERSION env var pins torch to a version the find-links mirror
   actually has — without pinning, pip sees newer PyPI CPU wheels and
   picks those over the CUDA variant (PEP 440 ordering: 2.11.0 >
   2.5.1+cu121 by major version). Defaults match pytorch.org so
   nothing changes for non-China runs.

Verified: pytest on Aliyun A10 + Qwen2.5-0.5B-Instruct (cached) —
6/6 PASS in 58s, max_abs_err == 0.0 across all 160 bit-exact
comparisons. Phase 1a pipeline validation GREEN on the small model.
Real Qwen2.5-3B acceptance run to follow.

+++ scripts/v6_replay/replay_engine.py
@@ -69,7 +69,7 @@ class ReplayEngine:
         )
         cache = out.past_key_values
         step_logits = out.logits[:, -1, :]
-        collected = [step_logits[target_index].detach().cpu().numpy()]
+        collected = [step_logits[target_index].detach().float().cpu().numpy()]
         next_tokens = torch.argmax(step_logits, dim=-1)
 
         for _ in range(1, batch_log.max_new_tokens):
@@ -92,7 +92,7 @@ class ReplayEngine:
             )
             cache = out.past_key_values
             step_logits = out.logits[:, -1, :]
-            collected.append(step_logits[target_index].detach().cpu().numpy())
+            collected.append(step_logits[target_index].detach().float().cpu().numpy())
             next_tokens = torch.argmax(step_logits, dim=-1)
 
         return TaskLogits(task_id=target_task_id, logits=collected)
